+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Adding `xan parallel --dont-chunk`.
 * Adding nullary `col_index` & `header()` variants, to work with expression applied in series to multiple columns at once.
 * Adding `xan filter -B/--before-context & -A/--after-context`.
+* Adding `xan window -O/--overwrite`.
 
 *Fixes*